requirements.txt

cloudinary==1.40.0
setuptools 
Django==4.1
django-allauth==0.60.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
django-widget-tweaks==1.4.12
google-generativeai      
gunicorn
psycopg2-binary==2.9.9
Pillow
django-ckeditor==6.4.2
django-filter